update(rename folder to their correct names)
